moved authorization middleware into subdirectories
