+++ lib/middleware/authorization/categories/category-access.ts
@@ -1,8 +1,8 @@
 import { NextFunction, Response } from 'express-serve-static-core';
-import { AuthenticatedReqParams } from '../../types/extendedExpress.js';
-import { Category } from '../../models/Category.js';
-import { List } from '../../models/List.js';
-import { ErrorWithStatus } from '../../types/error.js';
+import { AuthenticatedReqParams } from '../../../types/extendedExpress.js';
+import { Category } from '../../../models/Category.js';
+import { List } from '../../../models/List.js';
+import { ErrorWithStatus } from '../../../types/error.js';
 
 export default async (req: AuthenticatedReqParams<{ id: string }>, res: Response, next: NextFunction) => {
   try {
