+++ lib/middleware/authorization/list-shares/get-shared-users.ts
@@ -1,7 +1,7 @@
 import { NextFunction, Response } from 'express-serve-static-core';
-import { AuthenticatedRequest } from '../../types/extendedExpress.js';
-import { List } from '../../models/List.js';
-import { ErrorWithStatus } from '../../types/error.js';
+import { AuthenticatedRequest } from '../../../types/extendedExpress.js';
+import { List } from '../../../models/List.js';
+import { ErrorWithStatus } from '../../../types/error.js';
 
 export default async (req: AuthenticatedRequest, res: Response, next: NextFunction) => {
   try {
