+++ lib/middleware/authorization/list-shares/list-share.ts
@@ -1,8 +1,8 @@
 import { NextFunction, Response } from 'express-serve-static-core';
-import { AuthenticatedReqBody } from '../../types/extendedExpress.js';
-import { NewListShareData } from '../../types/listShare.js';
-import { List } from '../../models/List.js';
-import { ErrorWithStatus } from '../../types/error.js';
+import { AuthenticatedReqBody } from '../../../types/extendedExpress.js';
+import { NewListShareData } from '../../../types/listShare.js';
+import { List } from '../../../models/List.js';
+import { ErrorWithStatus } from '../../../types/error.js';
 
 export default async (req: AuthenticatedReqBody<NewListShareData>, res: Response, next: NextFunction) => {
   try {
