+++ lib/routes/lists.ts
@@ -2,7 +2,7 @@ import { Router } from 'express';
 import authenticate from '../middleware/authenticate.js';
 import { List, ListWithItems } from '../models/List.js';
 import { AuthenticatedReqBody, AuthenticatedReqParams, AuthenticatedRequest, TypedResponse } from '../types/extendedExpress.js';
-import authorizeListAccess from '../middleware/authorization/list-access.js';
+import authorizeListAccess from '../middleware/authorization/lists/list-access.js';
 import { ListRes, MultipleListsRes, NewListData } from '../types/list.js';
 import { NextFunction } from 'express-serve-static-core';
 import { Category } from '../models/Category.js';
